refactor: enhance the code structure and modularity
